+++ back/src/service/services/__init__.py
@@ -1,2 +1,3 @@
 from .usuario import criar_usuario, consultar_usuario, remover_usuario, atualizar_usuario
 from .servico import criar_servico, consultar_servico, listar_servicos, atualizar_servico, remover_servico
+from .jornada import criar_jornada, consultar_jornada, consultar_jornada_de_trabalho, alterar_ativacao_de_jornada, excluir_jornada
